con-rs: own bitmap helpers and missing-eb-tx logic
Moves the sparse BTreeMap<u16, u64> bitmap helpers
(from_indices / select_all / contains / iter_indices) into
con_rs::bitmap.  The encoded form already crosses the con-rs trait
boundary (EbTxsFetchPolicy::pick, on_eb_txs_offered), so its
encoders belong on the same side.  net-core's
leios_fetch::bitmap stays as a thin re-export to preserve callers.

Adds LeiosState::missing_eb_tx_bitmap(eb_hash, &MempoolState) →
the wire-encoded sparse bitmap of manifest indices whose bodies
aren't yet locally available.  Empty bitmap when the manifest is
unknown or every referenced tx is held — both deliberately
suppress further fetches (a "select all" reply against a peer
that doesn't have the bodies either triggers the retry storm we
hit on the broadcast-n=3 experiment).

Net-rs's bitmap_for_missing_txs now delegates to con-rs.
LeiosConsensus is the only consumer with both the LeiosState and
mempool handles, so the wrapper exposes a borrow via
Mempool::as_inner.  Dead Mempool::current_tx_ids method removed.
