+++ con-rs/src/lib.rs
@@ -7,6 +7,7 @@
 //! drive them with their own I/O layer.
 
 pub mod aggregation;
+pub mod bitmap;
 pub mod chain_tree;
 pub mod config;
 pub mod elections;
